// frontend/src/service.admin/work/types/WorkFormsControlsCollections.ts
interface WorkFormsControlsCollections extends HTMLFormControlsCollection {
	title: HTMLInputElement
	subject: HTMLInputElement
	description: HTMLTextAreaElement
	priority: HTMLInputElement
	active: HTMLInputElement
	youtubeLink: HTMLInputElement
	youtubeLinkName: HTMLInputElement
	externalLink: HTMLInputElement
	externalLinkName: HTMLInputElement
	isVideoPreviewImage: HTMLInputElement
}

export default WorkFormsControlsCollections